close #36 Fix controller
