+++ backend/app/adapter/http/controller/errors.go
@@ -1,5 +1,6 @@
 package controller
 
+var errMsgControllerPath = "controller path handling error"
 var errMsgWrongUserNameOrPassword = "Wrong username or password"
 var errMsgGuestUserForbidden = "not allowed to guest user"
 var errMsgUserNameResetKeyNotExistsResetKeyExpired = "the user name doesn't exist or the password reset key doesn't exist or the password reset key is expired"
