+++ backend/app/adapter/http/controller/follow_controller.go
@@ -5,8 +5,10 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"strings"
 
 	"github.com/gold-kou/ToeBeans/backend/app/lib"
+	"github.com/gorilla/mux"
 
 	"github.com/gold-kou/ToeBeans/backend/app/adapter/http/helper"
 	applicationLog "github.com/gold-kou/ToeBeans/backend/app/adapter/http/log"
@@ -23,26 +25,53 @@ func FollowController(w http.ResponseWriter, r *http.Request) {
 	}
 	l.LogHTTPAccess(r)
 
-	switch r.Method {
-	case http.MethodPost:
-		err = registerFollow(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.ForbiddenError:
-			helper.ResponseForbidden(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
+	switch {
+	case r.URL.Path == "/follow":
+		switch r.Method {
+		case http.MethodPost:
+			err = registerFollow(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.ForbiddenError:
+				helper.ResponseForbidden(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
 		default:
-			helper.ResponseInternalServerError(w, err.Error())
+			methods := []string{http.MethodPost}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
+		}
+	case strings.HasPrefix(r.URL.Path, "/follow/"):
+		switch r.Method {
+		case http.MethodDelete:
+			err = deleteFollow(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.ForbiddenError:
+				helper.ResponseForbidden(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
+		default:
+			methods := []string{http.MethodDelete}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
 		}
 	default:
-		methods := []string{http.MethodPost}
-		helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
+		helper.ResponseInternalServerError(w, errMsgControllerPath)
 	}
 }
 
@@ -102,6 +131,55 @@ func registerFollow(r *http.Request) error {
 		log.Println(err)
 		if err == repository.ErrDuplicateData {
 			return helper.NewBadRequestError("Whoops, you already followed the user")
+		}
+		return helper.NewInternalServerError(err.Error())
+	}
+	return err
+}
+
+func deleteFollow(r *http.Request) error {
+	// authorization
+	cookie, err := r.Cookie(helper.CookieIDToken)
+	if err != nil {
+		log.Println(err)
+		return helper.NewAuthorizationError(err.Error())
+	}
+	tokenUserName, err := lib.VerifyToken(cookie.Value)
+	if err != nil {
+		return helper.NewAuthorizationError(err.Error())
+	}
+	if tokenUserName == lib.GuestUserName {
+		log.Println(errMsgGuestUserForbidden)
+		return helper.NewForbiddenError(errMsgGuestUserForbidden)
+	}
+
+	// get request parameter
+	vars := mux.Vars(r)
+	followedUserName, ok := vars["followed_user_name"]
+	if !ok || followedUserName == "" {
+		log.Println(err)
+		return helper.NewBadRequestError("followed_user_name: cannot be blank.")
+	}
+
+	// db connect
+	db, err := mysql.NewDB()
+	if err != nil {
+		log.Println(err)
+		return helper.NewInternalServerError(err.Error())
+	}
+	defer db.Close()
+	tx := mysql.NewDBTransaction(db)
+
+	// repository
+	userRepo := repository.NewUserRepository(db)
+	followRepo := repository.NewFollowRepository(db)
+
+	// UseCase
+	u := usecase.NewDeleteFollow(r.Context(), tx, tokenUserName, followedUserName, userRepo, followRepo)
+	if err = u.DeleteFollowUseCase(); err != nil {
+		log.Println(err)
+		if err == repository.ErrNotExistsData {
+			return helper.NewBadRequestError(err.Error())
 		}
 		return helper.NewInternalServerError(err.Error())
 	}
