+++ backend/app/adapter/http/controller/like_controller.go
@@ -5,8 +5,11 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"strconv"
+	"strings"
 
 	"github.com/gold-kou/ToeBeans/backend/app/lib"
+	"github.com/gorilla/mux"
 
 	"github.com/gold-kou/ToeBeans/backend/app/adapter/http/helper"
 	applicationLog "github.com/gold-kou/ToeBeans/backend/app/adapter/http/log"
@@ -23,24 +26,49 @@ func LikeController(w http.ResponseWriter, r *http.Request) {
 	}
 	l.LogHTTPAccess(r)
 
-	switch r.Method {
-	case http.MethodPost:
-		err = registerLike(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
+	switch {
+	case r.URL.Path == "/like":
+		switch r.Method {
+		case http.MethodPost:
+			err = registerLike(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
 		default:
-			helper.ResponseInternalServerError(w, err.Error())
+			methods := []string{http.MethodPost}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
+		}
+	case strings.HasPrefix(r.URL.Path, "/like/"):
+		switch r.Method {
+		case http.MethodDelete:
+			err = deleteLike(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
+		default:
+			methods := []string{http.MethodDelete}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
 		}
 	default:
-		methods := []string{http.MethodPost}
-		helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
+		helper.ResponseInternalServerError(w, errMsgControllerPath)
 	}
 }
 
@@ -99,6 +127,56 @@ func registerLike(r *http.Request) error {
 			return helper.NewBadRequestError("Whoops, you already liked the posting")
 		}
 		if err == usecase.ErrLikeYourSelf {
+			return helper.NewBadRequestError(err.Error())
+		}
+		return helper.NewInternalServerError(err.Error())
+	}
+	return err
+}
+
+func deleteLike(r *http.Request) error {
+	// authorization
+	cookie, err := r.Cookie(helper.CookieIDToken)
+	if err != nil {
+		log.Println(err)
+		return helper.NewAuthorizationError(err.Error())
+	}
+	tokenUserName, err := lib.VerifyToken(cookie.Value)
+	if err != nil {
+		return helper.NewAuthorizationError(err.Error())
+	}
+
+	// get request parameter
+	vars := mux.Vars(r)
+	paramPostingID, ok := vars["posting_id"]
+	if !ok || paramPostingID == "0" {
+		log.Println(err)
+		return helper.NewBadRequestError("posting_id: cannot be blank")
+	}
+	postingID, err := strconv.Atoi(paramPostingID)
+	if err != nil {
+		return helper.NewInternalServerError(err.Error())
+	}
+
+	// db connect
+	db, err := mysql.NewDB()
+	if err != nil {
+		log.Println(err)
+		return helper.NewInternalServerError(err.Error())
+	}
+	defer db.Close()
+	tx := mysql.NewDBTransaction(db)
+
+	// repository
+	userRepo := repository.NewUserRepository(db)
+	postingRepo := repository.NewPostingRepository(db)
+	likeRepo := repository.NewLikeRepository(db)
+
+	// UseCase
+	u := usecase.NewDeleteLike(r.Context(), tx, tokenUserName, int64(postingID), userRepo, postingRepo, likeRepo)
+	if err = u.DeleteLikeUseCase(); err != nil {
+		log.Println(err)
+		if err == repository.ErrNotExistsData {
 			return helper.NewBadRequestError(err.Error())
 		}
 		return helper.NewInternalServerError(err.Error())
