--- backend/app/adapter/http/controller/like_posting_id_controller.go
@@ -1,95 +1,0 @@
-package controller
-
-import (
-	"log"
-	"net/http"
-	"strconv"
-
-	"github.com/gorilla/mux"
-
-	"github.com/gold-kou/ToeBeans/backend/app/lib"
-
-	"github.com/gold-kou/ToeBeans/backend/app/adapter/http/helper"
-	applicationLog "github.com/gold-kou/ToeBeans/backend/app/adapter/http/log"
-	"github.com/gold-kou/ToeBeans/backend/app/adapter/mysql"
-	"github.com/gold-kou/ToeBeans/backend/app/application/usecase"
-	"github.com/gold-kou/ToeBeans/backend/app/domain/repository"
-)
-
-func LikePostingIDController(w http.ResponseWriter, r *http.Request) {
-	l, err := applicationLog.NewLogger()
-	if err != nil {
-		log.Panic(err)
-	}
-	l.LogHTTPAccess(r)
-
-	switch r.Method {
-	case http.MethodDelete:
-		err = deleteLike(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
-		default:
-			helper.ResponseInternalServerError(w, err.Error())
-		}
-	default:
-		methods := []string{http.MethodDelete}
-		helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
-	}
-}
-
-func deleteLike(r *http.Request) error {
-	// authorization
-	cookie, err := r.Cookie(helper.CookieIDToken)
-	if err != nil {
-		log.Println(err)
-		return helper.NewAuthorizationError(err.Error())
-	}
-	tokenUserName, err := lib.VerifyToken(cookie.Value)
-	if err != nil {
-		return helper.NewAuthorizationError(err.Error())
-	}
-
-	// get request parameter
-	vars := mux.Vars(r)
-	paramPostingID, ok := vars["posting_id"]
-	if !ok || paramPostingID == "0" {
-		log.Println(err)
-		return helper.NewBadRequestError("posting_id: cannot be blank")
-	}
-	postingID, err := strconv.Atoi(paramPostingID)
-	if err != nil {
-		return helper.NewInternalServerError(err.Error())
-	}
-
-	// db connect
-	db, err := mysql.NewDB()
-	if err != nil {
-		log.Println(err)
-		return helper.NewInternalServerError(err.Error())
-	}
-	defer db.Close()
-	tx := mysql.NewDBTransaction(db)
-
-	// repository
-	userRepo := repository.NewUserRepository(db)
-	postingRepo := repository.NewPostingRepository(db)
-	likeRepo := repository.NewLikeRepository(db)
-
-	// UseCase
-	u := usecase.NewDeleteLike(r.Context(), tx, tokenUserName, int64(postingID), userRepo, postingRepo, likeRepo)
-	if err = u.DeleteLikeUseCase(); err != nil {
-		log.Println(err)
-		if err == repository.ErrNotExistsData {
-			return helper.NewBadRequestError(err.Error())
-		}
-		return helper.NewInternalServerError(err.Error())
-	}
-	return err
-}
