--- backend/app/adapter/http/controller/password_reset_controller.go
@@ -1,89 +1,0 @@
-package controller
-
-import (
-	"encoding/json"
-	"io/ioutil"
-	"log"
-	"net/http"
-
-	"github.com/gold-kou/ToeBeans/backend/app/adapter/mysql"
-	"github.com/gold-kou/ToeBeans/backend/app/domain/repository"
-
-	"github.com/gold-kou/ToeBeans/backend/app/application/usecase"
-
-	"github.com/gold-kou/ToeBeans/backend/app/adapter/http/helper"
-
-	applicationLog "github.com/gold-kou/ToeBeans/backend/app/adapter/http/log"
-	model "github.com/gold-kou/ToeBeans/backend/app/domain/model/http"
-)
-
-func PasswordResetController(w http.ResponseWriter, r *http.Request) {
-	l, err := applicationLog.NewLogger()
-	if err != nil {
-		log.Panic(err)
-	}
-	l.LogHTTPAccess(r)
-
-	switch r.Method {
-	case http.MethodPost:
-		err := passwordReset(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
-		default:
-			helper.ResponseInternalServerError(w, err.Error())
-		}
-	default:
-		methods := []string{http.MethodPost}
-		helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
-	}
-}
-
-func passwordReset(r *http.Request) error {
-	// get request parameter
-	var reqResetPassword *model.RequestResetPassword
-	b, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		log.Println(err)
-		return helper.NewBadRequestError(err.Error())
-	}
-	defer r.Body.Close()
-	if err = json.Unmarshal(b, &reqResetPassword); err != nil {
-		log.Println(err)
-		return helper.NewBadRequestError(err.Error())
-	}
-
-	// validation check
-	err = reqResetPassword.ValidateParam()
-	if err != nil {
-		log.Println(err)
-		return helper.NewBadRequestError(err.Error())
-	}
-
-	// db connect
-	db, err := mysql.NewDB()
-	if err != nil {
-		log.Println(err)
-		return helper.NewInternalServerError(err.Error())
-	}
-	defer db.Close()
-	tx := mysql.NewDBTransaction(db)
-
-	// repository
-	userRepo := repository.NewUserRepository(db)
-
-	// UseCase
-	re := usecase.NewPasswordReset(r.Context(), tx, reqResetPassword, userRepo)
-	if err = re.PasswordResetUseCase(); err != nil {
-		log.Println(err)
-		if err == repository.ErrNotExistsData {
-			return helper.NewBadRequestError(errMsgUserNameResetKeyNotExistsResetKeyExpired)
-		}
-		return helper.NewInternalServerError(err.Error())
-	}
-	return err
-}
