+++ backend/app/adapter/http/controller/posting_controller_test.go
@@ -2,11 +2,15 @@ package controller
 
 import (
 	"context"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/gorilla/mux"
 
 	"github.com/gold-kou/ToeBeans/backend/app/domain/repository"
 
@@ -186,6 +190,261 @@ func TestRegisterPosting(t *testing.T) {
 				postings[0].CreatedAt = lib.NowFunc()
 				postings[0].UpdatedAt = lib.NowFunc()
 				assert.Equal(t, dummy.Posting1, postings[0])
+			}
+
+			// assert http
+			assert.Equal(t, tt.wantStatus, resp.Code)
+			respBodyByte, err := ioutil.ReadAll(resp.Body)
+			assert.NoError(t, err)
+			respBody := string(respBodyByte)
+			assert.JSONEq(t, tt.want, respBody)
+		})
+	}
+}
+
+var successRespGetPostings = `
+{
+  "postings": [
+    {
+      "posting_id": 1,
+      "user_name": "testUser1",
+      "uploaded_at": "2020-01-01T00:00:00+09:00",
+      "title": "This is a sample posting.",
+      "image_url": "http://localhost:9000/postings/20200101000000_testUser1",
+      "liked_count": 0,
+      "liked": false
+    }
+  ]
+}
+`
+var errRespGetPostingsWithoutSinceAt = `
+{
+  "status": 400,
+  "message": "since_at: cannot be blank."
+}
+`
+var errRespGetPostingsWithoutLimit = `
+{
+  "status": 400,
+  "message": "limit: cannot be blank."
+}
+`
+var errRespGetPostingsNotExistsUserName = `
+{
+  "status": 400,
+  "message": "not exists data error"
+}
+`
+
+func TestGetPostings(t *testing.T) {
+	type args struct {
+		sinceAt  string
+		limit    string
+		userName string
+	}
+	tests := []struct {
+		name       string
+		args       args
+		method     string
+		want       string
+		wantStatus int
+	}{
+		{
+			name:       "success",
+			args:       args{sinceAt: "2020-01-01T00:00:00Z", limit: "50"},
+			method:     http.MethodGet,
+			want:       successRespGetPostings,
+			wantStatus: http.StatusOK,
+		},
+		{
+			name:       "success with user_name",
+			args:       args{sinceAt: "2020-01-01T00:00:00Z", limit: "50", userName: dummy.User1.Name},
+			method:     http.MethodGet,
+			want:       successRespGetPostings,
+			wantStatus: http.StatusOK,
+		},
+		{
+			name:       "error empty since_at",
+			args:       args{limit: "50"},
+			method:     http.MethodGet,
+			want:       errRespGetPostingsWithoutSinceAt,
+			wantStatus: http.StatusBadRequest,
+		},
+		{
+			name:       "error empty limit",
+			args:       args{sinceAt: "2020-01-01T00:00:00Z"},
+			method:     http.MethodGet,
+			want:       errRespGetPostingsWithoutLimit,
+			wantStatus: http.StatusBadRequest,
+		},
+		{
+			name:       "error not exists user_name",
+			args:       args{sinceAt: "2020-01-01T00:00:00Z", limit: "50", userName: dummy.User2.Name},
+			method:     http.MethodGet,
+			want:       errRespGetPostingsNotExistsUserName,
+			wantStatus: http.StatusBadRequest,
+		},
+		{
+			name:       "not allowed method",
+			args:       args{},
+			method:     http.MethodHead,
+			want:       testingHelper.ErrNotAllowedMethod,
+			wantStatus: http.StatusMethodNotAllowed,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// init
+			db := testingHelper.SetupDBTest()
+			defer testingHelper.TeardownDBTest(db)
+			testingHelper.SetTestTime()
+			defer testingHelper.ResetTime()
+
+			// insert dummy data
+			userRepo := repository.NewUserRepository(db)
+			postingRepo := repository.NewPostingRepository(db)
+			err := userRepo.Create(context.Background(), &dummy.User1)
+			assert.NoError(t, err)
+			err = postingRepo.Create(context.Background(), &dummy.Posting1)
+			assert.NoError(t, err)
+			err = testingHelper.UpdateNow(db, "postings")
+			assert.NoError(t, err)
+
+			// http request
+			var req *http.Request
+			if tt.args.userName == "" {
+				req, err = http.NewRequest(tt.method, fmt.Sprintf("/postings?since_at=%s&limit=%s", tt.args.sinceAt, tt.args.limit), nil)
+			} else {
+				req, err = http.NewRequest(tt.method, fmt.Sprintf("/postings?since_at=%s&limit=%s&user_name=%s", tt.args.sinceAt, tt.args.limit, tt.args.userName), nil)
+			}
+			assert.NoError(t, err)
+			token, err := lib.GenerateToken(dummy.User1.Name)
+			assert.NoError(t, err)
+			cookie := &http.Cookie{
+				Name:  helper.CookieIDToken,
+				Value: token,
+			}
+			req.AddCookie(cookie)
+			resp := httptest.NewRecorder()
+
+			// test target
+			PostingController(resp, req)
+			assert.NoError(t, err)
+
+			// assert http
+			assert.Equal(t, tt.wantStatus, resp.Code)
+			respBodyByte, err := ioutil.ReadAll(resp.Body)
+			assert.NoError(t, err)
+			respBody := string(respBodyByte)
+			assert.JSONEq(t, tt.want, respBody)
+		})
+	}
+}
+
+var errRespDeletePostingWithoutPostingID = `
+{
+  "status": 400,
+  "message": "posting_id: cannot be blank"
+}
+`
+var errRespDeletePostingNotExistingID = `
+{
+  "status": 400,
+  "message": "not exists data error"
+}
+`
+
+func TestDeletePosting(t *testing.T) {
+	type args struct {
+		postingID int64
+	}
+	tests := []struct {
+		name       string
+		args       args
+		method     string
+		want       string
+		wantStatus int
+	}{
+		{
+			name:       "success",
+			args:       args{postingID: dummy.Posting1.ID},
+			method:     http.MethodDelete,
+			want:       testingHelper.RespSimpleSuccess,
+			wantStatus: http.StatusOK,
+		},
+		{
+			name:       "error empty posting_id",
+			args:       args{},
+			method:     http.MethodDelete,
+			want:       errRespDeletePostingWithoutPostingID,
+			wantStatus: http.StatusBadRequest,
+		},
+		{
+			name:       "error not existing posting_id",
+			args:       args{postingID: 99999},
+			method:     http.MethodDelete,
+			want:       errRespDeletePostingNotExistingID,
+			wantStatus: http.StatusBadRequest,
+		},
+		{
+			name:       "error forbidden guest user",
+			args:       args{postingID: dummy.Posting1.ID},
+			method:     http.MethodDelete,
+			want:       testingHelper.ErrForbidden,
+			wantStatus: http.StatusForbidden,
+		},
+		{
+			name:       "not allowed method",
+			args:       args{},
+			method:     http.MethodHead,
+			want:       testingHelper.ErrNotAllowedMethod,
+			wantStatus: http.StatusMethodNotAllowed,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// init
+			db := testingHelper.SetupDBTest()
+			defer testingHelper.TeardownDBTest(db)
+			testingHelper.SetTestTime()
+			defer testingHelper.ResetTime()
+
+			// insert dummy data
+			userRepo := repository.NewUserRepository(db)
+			postingRepo := repository.NewPostingRepository(db)
+			err := userRepo.Create(context.Background(), &dummy.User1)
+			assert.NoError(t, err)
+			err = postingRepo.Create(context.Background(), &dummy.Posting1)
+			assert.NoError(t, err)
+
+			// http request
+			req, err := http.NewRequest(tt.method, fmt.Sprintf("/posting/%v", tt.args.postingID), nil)
+			assert.NoError(t, err)
+			vars := map[string]string{"posting_id": strconv.Itoa(int(tt.args.postingID))}
+			req = mux.SetURLVars(req, vars)
+			var token string
+			if tt.name == "error forbidden guest user" {
+				token, err = lib.GenerateToken(lib.GuestUserName)
+			} else {
+				token, err = lib.GenerateToken(dummy.User1.Name)
+			}
+			assert.NoError(t, err)
+			cookie := &http.Cookie{
+				Name:  helper.CookieIDToken,
+				Value: token,
+			}
+			req.AddCookie(cookie)
+			resp := httptest.NewRecorder()
+
+			// test target
+			PostingController(resp, req)
+			assert.NoError(t, err)
+
+			// assert db
+			if tt.wantStatus == 200 {
+				postings, err := testingHelper.FindAllPostings(context.Background(), db)
+				assert.NoError(t, err)
+				assert.Equal(t, 0, len(postings))
 			}
 
 			// assert http
