+++ backend/app/adapter/http/controller/user_controller.go
@@ -5,9 +5,11 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"strings"
 
 	validation "github.com/go-ozzo/ozzo-validation"
 	"github.com/go-ozzo/ozzo-validation/is"
+	"github.com/gorilla/mux"
 
 	"github.com/gold-kou/ToeBeans/backend/app/adapter/http/helper"
 	applicationLog "github.com/gold-kou/ToeBeans/backend/app/adapter/http/log"
@@ -26,85 +28,110 @@ func UserController(w http.ResponseWriter, r *http.Request) {
 	}
 	l.LogHTTPAccess(r)
 
-	switch r.Method {
-	case http.MethodPost:
-		err = registerUser(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
-		default:
-			helper.ResponseInternalServerError(w, err.Error())
-		}
-	case http.MethodGet:
-		user, err := getUser(r)
-		switch err := err.(type) {
-		case nil:
-			resp := modelHTTP.ResponseGetUser{
-				UserName:         user.Name,
-				Icon:             user.Icon,
-				SelfIntroduction: user.SelfIntroduction,
-				PostingCount:     user.PostingCount,
-				LikeCount:        user.LikeCount,
-				LikedCount:       user.LikedCount,
-				FollowCount:      user.FollowCount,
-				FollowedCount:    user.FollowedCount,
-				CreatedAt:        user.CreatedAt,
+	switch {
+	case r.URL.Path == "/user":
+		switch r.Method {
+		case http.MethodPost:
+			err = registerUser(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
 			}
-			w.Header().Set(helper.HeaderKeyContentType, helper.HeaderValueApplicationJSON)
-			w.WriteHeader(http.StatusOK)
-			if err := json.NewEncoder(w).Encode(resp); err != nil {
-				panic(err.Error())
+		case http.MethodGet:
+			user, err := getUser(r)
+			switch err := err.(type) {
+			case nil:
+				resp := modelHTTP.ResponseGetUser{
+					UserName:         user.Name,
+					Icon:             user.Icon,
+					SelfIntroduction: user.SelfIntroduction,
+					PostingCount:     user.PostingCount,
+					LikeCount:        user.LikeCount,
+					LikedCount:       user.LikedCount,
+					FollowCount:      user.FollowCount,
+					FollowedCount:    user.FollowedCount,
+					CreatedAt:        user.CreatedAt,
+				}
+				w.Header().Set(helper.HeaderKeyContentType, helper.HeaderValueApplicationJSON)
+				w.WriteHeader(http.StatusOK)
+				if err := json.NewEncoder(w).Encode(resp); err != nil {
+					panic(err.Error())
+				}
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.NotFoundError:
+				helper.ResponseNotFound(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
 			}
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.NotFoundError:
-			helper.ResponseNotFound(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
+		case http.MethodPut:
+			err = updateUser(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.ForbiddenError:
+				helper.ResponseForbidden(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
+		case http.MethodDelete:
+			err := deleteUser(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.ForbiddenError:
+				helper.ResponseForbidden(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
 		default:
-			helper.ResponseInternalServerError(w, err.Error())
+			methods := []string{http.MethodPost, http.MethodGet, http.MethodPut, http.MethodDelete}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
 		}
-	case http.MethodPut:
-		err = updateUser(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.ForbiddenError:
-			helper.ResponseForbidden(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
+	case strings.HasPrefix(r.URL.Path, "/user-activation/"):
+		switch r.Method {
+		case http.MethodGet:
+			err = activateUser(r)
+			switch err := err.(type) {
+			case nil:
+				helper.ResponseSimpleSuccess(w)
+			case *helper.BadRequestError:
+				helper.ResponseBadRequest(w, err.Error())
+			case *helper.AuthorizationError:
+				helper.ResponseUnauthorized(w, err.Error())
+			case *helper.InternalServerError:
+				helper.ResponseInternalServerError(w, err.Error())
+			default:
+				helper.ResponseInternalServerError(w, err.Error())
+			}
 		default:
-			helper.ResponseInternalServerError(w, err.Error())
-		}
-	case http.MethodDelete:
-		err := deleteUser(r)
-		switch err := err.(type) {
-		case nil:
-			helper.ResponseSimpleSuccess(w)
-		case *helper.BadRequestError:
-			helper.ResponseBadRequest(w, err.Error())
-		case *helper.AuthorizationError:
-			helper.ResponseUnauthorized(w, err.Error())
-		case *helper.ForbiddenError:
-			helper.ResponseForbidden(w, err.Error())
-		case *helper.InternalServerError:
-			helper.ResponseInternalServerError(w, err.Error())
-		default:
-			helper.ResponseInternalServerError(w, err.Error())
+			methods := []string{http.MethodGet}
+			helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
 		}
 	default:
-		methods := []string{http.MethodPost, http.MethodGet, http.MethodPut, http.MethodDelete}
-		helper.ResponseNotAllowedMethod(w, "not allowed method", methods)
+		helper.ResponseInternalServerError(w, errMsgControllerPath)
 	}
 }
 
@@ -314,6 +341,54 @@ func deleteUser(r *http.Request) (err error) {
 		log.Println(err)
 		if err == lib.ErrTokenInvalidNotExistingUserName {
 			return helper.NewAuthorizationError(err.Error())
+		}
+		return helper.NewInternalServerError(err.Error())
+	}
+	return
+}
+
+func activateUser(r *http.Request) (err error) {
+	// get path parameter
+	vars := mux.Vars(r)
+	userName, ok := vars["user_name"]
+	if !ok || userName == "" {
+		log.Println(err)
+		return helper.NewBadRequestError("user_name cannot be blank")
+	}
+	activationKey, ok := vars["activation_key"]
+	if !ok || activationKey == "" {
+		log.Println(err)
+		return helper.NewBadRequestError("activation_key cannot be blank")
+	}
+
+	// validation check
+	if err = validation.Validate(userName, validation.Required, validation.Length(modelHTTP.MinVarcharLength, modelHTTP.MaxVarcharLength), is.Alphanumeric); err != nil {
+		log.Println(err)
+		return helper.NewBadRequestError(err.Error())
+	}
+	if err = validation.Validate(activationKey, validation.Required, is.UUID); err != nil {
+		log.Println(err)
+		return helper.NewBadRequestError(err.Error())
+	}
+
+	// db connect
+	db, err := mysql.NewDB()
+	if err != nil {
+		log.Println(err)
+		return helper.NewInternalServerError(err.Error())
+	}
+	defer db.Close()
+	tx := mysql.NewDBTransaction(db)
+
+	// repository
+	userRepo := repository.NewUserRepository(db)
+
+	// UseCase
+	u := usecase.NewUserActivation(r.Context(), tx, userName, activationKey, userRepo)
+	if err = u.UserActivationUseCase(); err != nil {
+		log.Println(err)
+		if err == repository.ErrUserActivationNotFound {
+			return helper.NewBadRequestError(err.Error())
 		}
 		return helper.NewInternalServerError(err.Error())
 	}
